+++ src/Pages/Login/Login.js
@@ -1,5 +1,118 @@
-import React from 'react';
-import { LoginForm } from '../../Components/Login/LoginForm';
+import React, { useState } from 'react';
+import { useAuth } from '../../Components/context/AuthContext';
+import { loginUser } from '../../APIFunctions/Auth';
+
+const LoginForm = ({ className, ...props }) => {
+  const { setAuthenticated } = useAuth();
+  const [errorMsg, setErrorMsg] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+  const queryParams = new URLSearchParams(window.location.search);
+
+  async function handleSubmit(e) {
+    e.preventDefault();
+    setIsLoading(true);
+    setErrorMsg('');
+
+    const formData = new FormData(e.target);
+    const email = formData.get('email');
+    const password = formData.get('password');
+
+    try {
+      const loginStatus = await loginUser(email, password);
+      if (!loginStatus.error) {
+        setAuthenticated(true);
+        window.localStorage.setItem('jwtToken', loginStatus.token);
+        if (queryParams.get('redirect')) {
+          window.location.href = queryParams.get('redirect');
+        } else {
+          window.location.reload();
+        }
+      } else {
+        if (loginStatus.responseData === undefined || loginStatus.responseData.data.message === undefined) {
+          setErrorMsg('Backend May Be down, check with dev team!');
+        } else {
+          setErrorMsg(loginStatus.responseData && loginStatus.responseData.data.message);
+        }
+      }
+    } catch (error) {
+      setErrorMsg('An unexpected error occurred. Please try again.');
+    } finally {
+      setIsLoading(false);
+    }
+  }
+
+  return (
+    <div className={`flex flex-col gap-6 ${className || ''}`} {...props}>
+      <div className="card bg-base-100 border border-white/10">
+        <div className="card-body">
+          <div className="text-center mb-6">
+            <h2 className="card-title text-2xl font-bold justify-center">Login to your account</h2>
+            <p className="text-base-content/70 text-sm mt-2">
+              Enter your email below to login to your account
+            </p>
+          </div>
+          <form onSubmit={handleSubmit}>
+            <div className="flex flex-col gap-6">
+              <div className="grid gap-3">
+                <label htmlFor="email" className="label">
+                  <span className="label-text text-sm font-medium">Email</span>
+                </label>
+                <input
+                  id="email"
+                  name="email"
+                  type="email"
+                  placeholder="m@example.com"
+                  required
+                  className="input input-bordered w-full"
+                />
+              </div>
+              <div className="grid gap-3">
+                <div className="flex items-center">
+                  <label htmlFor="password" className="label">
+                    <span className="label-text text-sm font-medium">Password</span>
+                  </label>
+                  <a
+                    href="/forgot"
+                    className="ml-auto text-sm underline-offset-4 hover:underline"
+                  >
+                    Forgot your password?
+                  </a>
+                </div>
+                <input
+                  id="password"
+                  name="password"
+                  type="password"
+                  required
+                  className="input input-bordered w-full"
+                />
+              </div>
+              {errorMsg && (
+                <div className="text-red-500 text-sm text-center">
+                  {errorMsg}
+                </div>
+              )}
+              <div className="flex flex-col gap-3">
+                <button
+                  type="submit"
+                  className="btn btn-primary w-full"
+                  disabled={isLoading}
+                >
+                  {isLoading ? 'Logging in...' : 'Login'}
+                </button>
+              </div>
+            </div>
+            <div className="mt-4 text-center text-sm">
+              Don&apos;t have an account?{' '}
+              <a href="/register" className="underline underline-offset-4">
+                Sign up
+              </a>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  );
+};
 
 export default function Login() {
   return (
